style refactor

+++ src/main/python/utime.py
@@ -20,7 +20,6 @@ def localtime(secs) -> tuple:
 
     http://docs.micropython.org/en/latest/library/utime.html#utime.localtime
     """
-    
 
 def mktime():
     """
@@ -125,7 +124,7 @@ def ticks_us():
     """
     
 
-def time() -> int:
+def time() -> float:
     """
     Returns the number of seconds, as an integer, since the Epoch, assuming that underlying RTC is set and maintained as described above. If an RTC is not set, this function returns number of seconds since a port-specific reference point in time (for embedded boards without a battery-backed RTC, usually since power up or reset). If you want to develop portable MicroPython application, you should not rely on this function to provide higher than second precision. If you need higher precision, use ticks_ms() and ticks_us() functions, if you need calendar time, localtime() without an argument is a better choice.
